# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مقرأة سَنا الآي Quran Center Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/login/select:
    post:
      operationId: loginSelectAccount
      tags: [auth]
      summary: Select account after multi-role login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginSelectBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
components:
...
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginSelectBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        accountId:
          type: integer
      required:
        - email
        - password
        - accountId
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        requiresSelection:
          type: boolean
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountOption"
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [leader, data_entry, teacher, supervisor, student, track_supervisor, volunteer, exam_supervisor]
        track:
          type: ["string", "null"]
        circleId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        ageRange:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        isArchived:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - isArchived
        - createdAt
...
    AccountOption:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        track:
          type: ["string", "null"]
        circleId:
          type: ["integer", "null"]
      required:
        - id
        - name
        - role
...